Removes duplicated specs
