+++ spec/support/configs/simple_cov_config.rb
@@ -4,7 +4,8 @@ require 'simplecov_json_formatter'
 
 module SimpleCovConfig
   def self.configure
-    SimpleCov.formatter = SimpleCov::Formatter::JSONFormatter
+    SimpleCov.formatter = SimpleCov::Formatter::JSONFormatter if ENV['CC_TEST_REPORTER_ID']
+
     SimpleCov.minimum_coverage 90
     SimpleCov.start do
       add_filter { |source_file| cover?(source_file.lines) }
